Website/APIs_Docker/classify/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
opencv-python
inference-sdk
torch